Add: changed noteroutes to controller.js and beginning refactor to async/await
