+++ server.js
@@ -5,7 +5,6 @@ const cors = require('cors');
 
 // bring in the routes
 const routes = require('./api/routes');
-const notesRoutes = require('./api/notes/notesRoutes');
 
 // create a server
 const server = express();
@@ -23,7 +22,6 @@ const corsOptions = {
 };
 
 server.use(cors(corsOptions));
-server.use('/api/notes/notesRoutes', notesRoutes);
 
 // pass the server to the routes
 routes(server);
